Rename profile-related stuff
