+++ react/locales/pt-BR.json
@@ -3,7 +3,7 @@
 
   "pages.orders": "Meus pedidos",
   "pages.addresses": "Endereços",
-  "pages.personalData": "Dados pessoais",
+  "pages.profile": "Dados pessoais",
   "pages.paymentData": "Métodos de pagamento",
   "pages.myAccount": "Minha conta",
